use NSUInteger for hash
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@16468 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKFile.m
@@ -51,7 +51,7 @@
 @interface BDSKFSRefFile : BDSKFile <NSCopying>
 {
     const FSRef *fileRef;
-    UInt32 hash;
+    NSUInteger hash;
 }
 @end
